Special/Msphere: Abstraction w.i.p.
